.ipynb_checkpoints/requirements-checkpoint.txt
aiofiles==24.1.0
annotated-doc==0.0.4
audioread==3.1.0
catalogue==2.0.10
certifi==2025.11.12
cfgv==3.4.0
charset-normalizer==3.4.4
chatterbox-tts==0.1.4
conformer==0.3.2
Deprecated==1.3.1
diffusers==0.29.0
distlib==0.4.0
einops==0.8.1
fastapi==0.121.1
ffmpy==1.0.0
filelock==3.20.0
flash_attn==2.8.3
fsspec==2025.10.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
hf-xet==1.2.0
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
jaconv==0.4.0
librosa==0.11.0
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.44.0
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
munkres==1.1.4
nodeenv==1.9.1
numpy==1.26.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.19.1
orjson==3.11.4
packaging==25.0
pooch==1.8.2
pre_commit==4.4.0
pyarrow==21.0.0
pydub==0.25.1
pykakasi==2.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.5
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resemble-perth==1.0.1
rich==14.2.0
ruff==0.14.4
s3tokenizer==0.2.0
safehttpx==0.1.7
safetensors==0.6.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
soundfile==0.13.1
soxr==1.0.0
spacy_pkuseg==1.0.1
srsly==2.5.1
starlette==0.49.3
sympy==1.14.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.9.0
torchaudio==2.9.0
torchmetrics==1.8.2
torchvision==0.24.0
tqdm==4.67.1
transformers==4.51.1
triton==3.5.0
typer==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
websockets==15.0.1
wheel==0.45.1
wrapt==2.0.1
zstandard==0.24.0
olaph